Separated Matrix and operations into header files
