+++ .gitignore
@@ -8,3 +8,5 @@ scratch.cpp
 package_tests
 .vscode/c_cpp_properties.json
 .vscode/launch.json
+*.ppm
+*.o
