+++ .vscode/settings.json
@@ -13,6 +13,7 @@
         "cmath": "cpp",
         "complex": "cpp",
         "chrono": "cpp",
-        "string": "cpp"
+        "string": "cpp",
+        "random": "cpp"
     }
 }
