+++ examples/matrix_inverse/matrix_inverse.cpp
@@ -1,0 +1,7 @@
+/*
+This example compares the accuracy of ill-conditioned matrix inversion 
+in different number systems: float and posit32.
+*/
+
+#include <posit_ops.h>
+#include <Matrix.h>
